+++ spec/configuration_spec.rb
@@ -33,5 +33,29 @@ describe SimpleCov::Configuration do
         expect(config.tracked_files).to be_nil
       end
     end
+
+    describe "#minimum_coverage" do
+      it "does not warn you about your usage" do
+        expect(config).not_to receive(:warn)
+        config.minimum_coverage(100.00)
+      end
+
+      it "warns you about your usage" do
+        expect(config).to receive(:warn).with("The coverage you set for minimum_coverage is greater than 100%")
+        config.minimum_coverage(100.01)
+      end
+    end
+
+    describe "minimum_coverage_by_file" do
+      it "does not warn you about your usage" do
+        expect(config).not_to receive(:warn)
+        config.minimum_coverage_by_file(100.00)
+      end
+
+      it "warns you about your usage" do
+        expect(config).to receive(:warn).with("The coverage you set for minimum_coverage_by_file is greater than 100%")
+        config.minimum_coverage_by_file(100.01)
+      end
+    end
   end
 end